Refactored ViewableLogic
